Move CLI commands into separate funcs
